Change Descriptive module to use StateT
@nikita-volkov Here's a first stab on the `monadic` branch. It generally
simplifies the code. The other modules under Descriptive.* will have to
wait. It's mostly a mechanical translation, but I've not much more time
this evening.

It's nice that I'm able to preserve the types of consume/describe and
that the particular `m` in question doesn't have to be added to the
Consumer type but rather is decided when calling
runConsumer/runDescription. Good old Rank-N types. What would we do
without them?